Fix s3d::InRange argument type mismatch in Game.cpp
Corrected calls to the s3d::InRange function within
Game::GenerateAndSetupNewMap to ensure argument type consistency.
The third argument (max bound, derived from currentMapGrid.width()
and .height()) is now explicitly cast to int using static_cast<int>().

This resolves the "no instance of overloaded function InRange matches
the argument list" error, which was likely caused by mixing int and
size_t types in the function call.

+++ Game.cpp
@@ -111,8 +111,8 @@ void Game::GenerateAndSetupNewMap() {
 
                 // Check if the randomly chosen position is within the map grid bounds
                 // AND is a floor tile (type 0) in currentMapGrid.
-                if (s3d::InRange(spawnPos.x, 0, currentMapGrid.width() - 1) &&
-                    s3d::InRange(spawnPos.y, 0, currentMapGrid.height() - 1) &&
+                if (s3d::InRange(spawnPos.x, 0, static_cast<int>(currentMapGrid.width() - 1)) &&
+                    s3d::InRange(spawnPos.y, 0, static_cast<int>(currentMapGrid.height() - 1)) &&
                     currentMapGrid[spawnPos.y][spawnPos.x] == 0) {
 
                     Enemys << new BaseEnemy(spawnPos, 0); // Create and add new enemy (type 0)
